Feed tb_cpe_versions database if not fully aligned
When not all versions in the tb_cpe_?_* tables are available in the
tb_cpe_versions table, then cvecheckers' ability to present vulnerabilities in
other (newer) versions versus existing version.

Introduce the feed_cpe_versions_table() that fills this table if it is not full,
and call it either when the tb_cpe_versions table doesn't exist yet, or when the
counts do not match.