Fixed missing increment in line number in case of a defect line
The removed lines of code were unnecessary, and resulted in invalid lines
reported for consecutive errors as the linenum variable is not increased.

The removed lines of code are creating that bug, because they continue the loop
before increasing the linenum variable. Since the fixed code before the
"continue" would be the same as the current end of the loop body just fall
through to that code instead of duplicating it.