Switch to JSON-based CVE feed
The NVD site no longer provides XML-based CVE feeds. Instead, a JSON
format is used. So we switch from XML processing to JSON.

To be able to handle this efficiently, we add in dependency on the jq
application, which is a powerful command-line JSON parser. This
eliminates the need for the XSL processing from XML to CSV.

This fixes #46.